+++ packages/atomic/cypress/integration/smart-snippet-selectors.ts
@@ -4,13 +4,56 @@ export const SmartSnippetSelectors = {
   accessibilityHeading: () =>
     SmartSnippetSelectors.shadow().find('.accessibility-only'),
   smartSnippet: () =>
-    SmartSnippetSelectors.shadow().find('[part="smart-snippet"]').first(),
-  question: () =>
-    SmartSnippetSelectors.shadow().find('[part="question"]').first(),
-  answer: () => SmartSnippetSelectors.shadow().find('[part="answer"]').first(),
-  sourceUrl: () =>
-    SmartSnippetSelectors.shadow().find('[part="source-url"] a').first(),
+    SmartSnippetSelectors.shadow().find('[part="smart-snippet"]'),
+  question: () => SmartSnippetSelectors.shadow().find('[part="question"]'),
+  answer: () =>
+    SmartSnippetSelectors.shadow().find('[part="answer"]', {
+      includeShadowDom: true,
+    }),
+  truncatedAnswer: () =>
+    SmartSnippetSelectors.shadow().find('[part="truncated-answer"]', {
+      includeShadowDom: true,
+    }),
+  showMoreButton: () =>
+    SmartSnippetSelectors.shadow().find('[part="show-more-button"]', {
+      includeShadowDom: true,
+    }),
+  showLessButton: () =>
+    SmartSnippetSelectors.shadow().find('[part="show-less-button"]', {
+      includeShadowDom: true,
+    }),
+  body: () => SmartSnippetSelectors.shadow().find('[part="body"]'),
+  sourceUrl: () => SmartSnippetSelectors.shadow().find('[part="source-url"] a'),
   sourceTitle: () =>
-    SmartSnippetSelectors.shadow().find('[part="source-title"] a').first(),
-  footer: () => SmartSnippetSelectors.shadow().find('[part="footer"]').first(),
+    SmartSnippetSelectors.shadow().find('[part="source-title"] a'),
+  footer: () => SmartSnippetSelectors.shadow().find('[part="footer"]'),
+  feedbackBanner: () =>
+    SmartSnippetSelectors.shadow().find('[part="feedback-banner"]', {
+      includeShadowDom: true,
+    }),
+  feedbackInquiryAndButtons: () =>
+    SmartSnippetSelectors.shadow().find(
+      '[part="feedback-inquiry-and-buttons"]',
+      {includeShadowDom: true}
+    ),
+  feedbackInquiry: () =>
+    SmartSnippetSelectors.shadow().find('[part="feedback-inquiry"]', {
+      includeShadowDom: true,
+    }),
+  feedbackButtons: () =>
+    SmartSnippetSelectors.shadow().find('[part="feedback-buttons"]', {
+      includeShadowDom: true,
+    }),
+  feedbackLikeButton: () =>
+    SmartSnippetSelectors.shadow().find('[part="feedback-like-button"]', {
+      includeShadowDom: true,
+    }),
+  feedbackDislikeButton: () =>
+    SmartSnippetSelectors.shadow().find('[part="feedback-dislike-button"]', {
+      includeShadowDom: true,
+    }),
+  feedbackThankYou: () =>
+    SmartSnippetSelectors.shadow().find('[part="feedback-thank-you"]', {
+      includeShadowDom: true,
+    }),
 };