+++ packages/atomic/src/utils/utils.ts
@@ -34,12 +34,13 @@ export function titleToKebab(value: string) {
 }
 
 export function randomID(prepend?: string, length = 5) {
-  return (
-    prepend +
-    Math.random()
-      .toString(36)
-      .substr(2, 2 + length)
-  );
+  const randomStr = Math.random()
+    .toString(36)
+    .substring(2, 2 + length);
+  if (!prepend) {
+    return randomStr;
+  }
+  return prepend + randomStr;
 }
 
 export function getRandomArbitrary(min: number, max: number) {