+++ packages/headless/src/features/question-answering/question-answering-slice.ts
@@ -1,5 +1,9 @@
 import {createReducer} from '@reduxjs/toolkit';
-import {QuestionAnswerDocumentIdentifier} from '../../api/search/search/question-answering';
+import {
+  QuestionAnswer,
+  QuestionAnswerDocumentIdentifier,
+} from '../../api/search/search/question-answering';
+import {encodedBtoa} from '../../utils/utils';
 import {executeSearch} from '../search/search-actions';
 import {
   collapseSmartSnippet,
@@ -25,6 +29,21 @@ export const findRelatedQuestionIdx = (
       relatedQuestion.contentIdValue === identifier.contentIdValue &&
       relatedQuestion.contentIdKey === identifier.contentIdKey
   );
+
+function hashQuestionAnswer({
+  question,
+  answerSnippet,
+  documentId: {contentIdKey, contentIdValue},
+}: QuestionAnswer) {
+  return encodedBtoa(
+    JSON.stringify({
+      question,
+      answerSnippet,
+      contentIdKey,
+      contentIdValue,
+    })
+  );
+}
 
 export const questionAnsweringReducer = createReducer(
   getQuestionAnsweringInitialState(),
@@ -52,7 +71,7 @@ export const questionAnsweringReducer = createReducer(
         state.feedbackModalOpened = false;
       })
       .addCase(executeSearch.fulfilled, (state, action) => {
-        state.relatedQuestions =
+        const relatedQuestions =
           action.payload.response.questionAnswer.relatedQuestions.map(
             (relatedQuestion) => ({
               contentIdKey: relatedQuestion.documentId.contentIdKey,
@@ -60,6 +79,20 @@ export const questionAnsweringReducer = createReducer(
               expanded: false,
             })
           );
+        const questionAnswerId = hashQuestionAnswer(
+          action.payload.response.questionAnswer
+        );
+        if (state.questionAnswerId === questionAnswerId) {
+          return {
+            ...state,
+            relatedQuestions,
+          };
+        }
+        return {
+          ...getQuestionAnsweringInitialState(),
+          relatedQuestions,
+          questionAnswerId,
+        };
       })
       .addCase(expandSmartSnippetRelatedQuestion, (state, action) => {
         const idx = findRelatedQuestionIdx(