+++ packages/atomic-angular/projects/atomic-angular/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.7.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@0.6.4...@coveo/atomic-angular@0.7.0) (2022-03-25)
+
+
+### Features
+
+* **atomic:** added smart snippet expandable answer ([#1841](https://github.com/coveo/ui-kit/issues/1841)) ([d7169d9](https://github.com/coveo/ui-kit/commit/d7169d96abacecbccdcc9accc9199688f44cdd31))
+
+
+
+
+
+## [0.6.4](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@0.6.3...@coveo/atomic-angular@0.6.4) (2022-03-24)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [0.6.3](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@0.6.2...@coveo/atomic-angular@0.6.3) (2022-03-24)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
 ## [0.6.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@0.6.1...@coveo/atomic-angular@0.6.2) (2022-03-22)
 
 **Note:** Version bump only for package @coveo/atomic-angular