+++ packages/samples/headless-react/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.51.3](https://github.com/coveo/ui-kit/compare/headless-react@0.51.2...headless-react@0.51.3) (2022-03-25)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.51.2](https://github.com/coveo/ui-kit/compare/headless-react@0.51.1...headless-react@0.51.2) (2022-03-24)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.51.1](https://github.com/coveo/ui-kit/compare/headless-react@0.51.0...headless-react@0.51.1) (2022-03-24)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
 # [0.51.0](https://github.com/coveo/ui-kit/compare/headless-react@0.50.38...headless-react@0.51.0) (2022-03-22)